Move the market items sorting logic from DataSource to ViewModel
