+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/top/MarketListTopDataSource.kt
@@ -6,7 +6,7 @@ import io.horizontalsystems.xrateskit.entities.CoinMarket
 import io.reactivex.Observable
 import io.reactivex.Single
 
-class MarketListTopDataSource(private val xRateManager: IRateManager) : IMarketListDataSource() {
+class MarketListTopDataSource(private val xRateManager: IRateManager) : IMarketListDataSource {
 
     override val sortingFields: Array<Field> = arrayOf(
             Field.HighestCap,
@@ -21,7 +21,7 @@ class MarketListTopDataSource(private val xRateManager: IRateManager) : IMarketL
 
     override val dataUpdatedAsync: Observable<Unit> = Observable.empty()
 
-    override fun doGetListAsync(currencyCode: String, fetchDiffPeriod: TimePeriod): Single<List<CoinMarket>> {
+    override fun getListAsync(currencyCode: String, fetchDiffPeriod: TimePeriod): Single<List<CoinMarket>> {
         return xRateManager.getTopMarketList(currencyCode, fetchDiffPeriod)
     }
 }
